// src/types/CalendarEvent.d.ts
export interface CalendarEvent {
  id: number;
  title: string;
  date: Date;
}
export interface series {
  id: number;
  seriesName: string;
}
export interface sponsor {
  id: number;
  sponsorName: string;
}
export interface location {
  id: number;
  locationName: string;
}
export interface division {
  id: number;
  divisionName: string;
}
interface filteredData {
  id: number;
  name: string;
}